# Speed up evaluation with r-trees to find overlapping detections