--- a/fiftyone/utils/eval/coco.py
+++ b/fiftyone/utils/eval/coco.py
@@ -527,12 +527,9 @@ def _coco_evaluation_setup(
         # Sort ground truth so crowds are last
         gts = sorted(gts, key=iscrowd)
 
-        # Compute ``num_preds x num_gts`` IoUs
+        # Compute IoUs for overlapping detections
         ious = foui.compute_ious(preds, gts, **iou_kwargs)
-
-        gt_ids = [g.id for g in gts]
-        for pred, gt_ious in zip(preds, ious):
-            pred_ious[pred.id] = list(zip(gt_ids, gt_ious))
+        pred_ious.update(ious)
 
     return cats, pred_ious, iscrowd