## What changes are proposed in this pull request?

The current evaluation methods generate IOUs for all pairs of detections, which is quadratic in both time and space. It's taking hours even though most detections don't overlap at all. This PR uses a spatial index (implemented in the `rtree` package) to only compute IOUs for detections that overlap. It takes the run time from hours to minutes on images with tens of thousands of detections.

I ran a simple test with 10,000 detections. This change improved the run time from 945 seconds to 9.5 seconds.

It works for 2D and 3D bounding boxes, masked detections, and polygons. Keypoints and polylines are not changed at this time, except to update the return value of `compute_ious()` from an array to a dictionary, to match the others.

This PR also fixes a bug in `utils3d._Box`, which was using the wrong rotation to compute the box's vertices. This also allows the cuboid IOU tests to pass.

Note that providing an `eval_key` to save the results to the database is unreasonably slow, likely because of the overhead of saving the entire sample. That problem is unrelated to the detection matching issues solved here, but it should be addressed in another PR. For instance, in a test with 1,000 detections, the runtime with `eval_key=None` goes from 9.3 seconds to 0.8 seconds. However, with an `eval_key`, it is about 65 seconds in both cases.

## How is this patch tested? If it is not, please explain why.

`evaluation_tests.py` confirms that the new results match the old ones. The `CuboidTests` class had to be updated because now the computed ious do not include results for non-overlapping detections.

## Release Notes

### Is this a user-facing change that should be mentioned in the release notes?

-   [ ] No. You can skip the rest of this section.
-   [X] Yes. Give a description of this change to be included in the release
        notes for FiftyOne users.

It does speed up evaluations, so users with larger numbers of objects in their samples may be interested.

### What areas of FiftyOne does this PR affect?

-   [ ] App: FiftyOne application changes
-   [ ] Build: Build and test infrastructure changes
-   [X] Core: Core `fiftyone` Python library changes
-   [ ] Documentation: FiftyOne documentation changes
-   [ ] Other

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
	- Enhanced handling of Intersection over Union (IoU) calculations for improved performance and clarity.
	- Introduced new functions for structured extraction of bounding box coordinates.
	- Added support for spatial indexing with the integration of the "rtree" library.

- **Bug Fixes**
	- Improved validation logic for IoU calculations in unit tests.

- **Tests**
	- Added new test class and method to enhance coverage for box functionality in 3D utilities.

- **Chores**
	- Updated dependencies to include "rtree" for enhanced spatial operations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->