@swheaton 2024-09-03
Thanks for your contribution @kemaleren ! Someone will review your proposal.

Re: saving results, we have been looking into that separately. Stay tuned.